src/finetune/pyproject.toml
[project]
name = "ovd-finetune"
version = "0.1.0"
description = "OVD Platform — Pipeline de fine-tuning y generación de datos sintéticos"
requires-python = ">=3.12"
dependencies = [
    # Base — exportación, validación, benchmark, generación sintética
    "psycopg[binary]>=3.1",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
# GPU — fine-tuning OSS local (requiere CUDA, VRAM mínima: 16GB para 7B)
# Instalar con: uv sync --extra gpu
gpu = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "sentencepiece>=0.2.0",
    "tokenizers>=0.19.0",
]

[tool.uv]
# Uso:
#   uv sync              → solo dependencias base (export, validate, benchmark)
#   uv sync --extra gpu  → incluye Unsloth/torch para fine-tuning local con GPU
